refactor : localStorage 사용 파일 이름 변경 및 적용
